analytics-etl/requirements.txt
pandas
clickhouse-connect
psycopg2-binary
